Sami sprint5

+++ Frontend/channelx/src/components/Home/Home.js
@@ -466,24 +466,28 @@ class Home extends Component {
     deleteExpiredChannels = () => {
         const messagingChannelDeleter = new MessagingChannelDeleter();
         console.log("Inside DeleteExpiredChannel");
-        var time;
         var today = new Date(),
             date = (today.getMonth() + 1) + '/' + today.getDate() + '/' + today.getFullYear();
-        time = Moment(today).format('HH:mm:ss').toString();
 
-        console.log(date);
-        db.collection("channels").where("channelEndDate", "<", date)
-            .get()
-            .then(snapshot => {
-                snapshot
-                    .docs
-                    .forEach(doc => {
-                        //console.log("Expired Channel id: " + doc.id);
-                        doc.ref.delete();
-                        messagingChannelDeleter.deleteChannel(doc.id);
-                    })
+        db.collection("channels").get().then(ref => {
+            ref.docs.forEach(doc => {
+
+                console.log(doc.id);
+
+                var channelEndDate = doc.get("channelEndDate");
+                var validDate = Moment(date).isSameOrBefore(channelEndDate);
+
+                console.log(validDate);
+
+                if(!validDate){
+                    doc.ref.delete();
+                    messagingChannelDeleter.deleteChannel(doc.id);
+                }
+
             });
-    };
+        });
+
+    }
 
     deleteChannelClicked = (channelTitle) => {
         const messagingChannelDeleter = new MessagingChannelDeleter();
